10/10/25
Started a bit of general cleanup and also made progress with enemy selection.

+++ Assets/_Dev/CHauser/Scripts/Actor.cs
@@ -11,7 +11,7 @@ public class Actor : MonoBehaviour
     [SerializeField] public float health;
     [SerializeField] public float maxHealth;
     [SerializeField] public float speed;
-    [SerializeField] public float range; // Range is the number of gird squares away an actor can attack.
+    [SerializeField] public int range; // Range is the number of gird squares away an actor can attack.
 
 
     public Agent agent;
